Extract module `kontrol.kompile`
