+++ src/kontrol/__main__.py
@@ -16,7 +16,6 @@ from .cli import KontrolCLIArgs
 from .foundry import (
     Foundry,
     foundry_get_model,
-    foundry_kompile,
     foundry_list,
     foundry_merge_nodes,
     foundry_node_printer,
@@ -27,6 +26,7 @@ from .foundry import (
     foundry_step_node,
     foundry_to_dot,
 )
+from .kompile import foundry_kompile
 from .prove import foundry_prove
 from .solc_to_k import solc_compile, solc_to_k
