+++ src/kontrol/foundry.py
@@ -4,7 +4,6 @@ import json
 import logging
 import os
 import re
-import shutil
 import sys
 from functools import cached_property
 from os import listdir
@@ -13,12 +12,10 @@ from subprocess import CalledProcessError
 from typing import TYPE_CHECKING
 
 import tomlkit
-from kevm_pyk.dist import DistTarget
 from kevm_pyk.kevm import KEVM, KEVMNodePrinter, KEVMSemantics
 from kevm_pyk.utils import byte_offset_to_lines, legacy_explore, print_failure_info, print_model
 from pyk.kast.inner import KApply, KSort, KToken
 from pyk.kast.manip import minimize_term
-from pyk.kast.outer import KDefinition, KFlatModule, KImport, KRequire
 from pyk.kcfg import KCFG
 from pyk.prelude.bytes import bytesToken
 from pyk.prelude.kbool import notBool
@@ -28,7 +25,7 @@ from pyk.proof.reachability import APRBMCProof, APRProof
 from pyk.proof.show import APRBMCProofNodePrinter, APRProofNodePrinter, APRProofShow
 from pyk.utils import ensure_dir_path, hash_str, run_process, single, unique
 
-from .solc_to_k import Contract, contract_to_main_module, contract_to_verification_module
+from .solc_to_k import Contract
 
 if TYPE_CHECKING:
     from collections.abc import Iterable
@@ -40,6 +37,7 @@ if TYPE_CHECKING:
     from pyk.kcfg.tui import KCFGElem
     from pyk.proof.show import NodePrinter
     from pyk.utils import BugReport
+
 
 _LOGGER: Final = logging.getLogger(__name__)
 
@@ -461,132 +459,6 @@ class Foundry:
         return latest_version + 1 if latest_version is not None else 0
 
 
-def foundry_kompile(
-    foundry_root: Path,
-    includes: Iterable[str],
-    regen: bool = False,
-    rekompile: bool = False,
-    requires: Iterable[str] = (),
-    imports: Iterable[str] = (),
-    ccopts: Iterable[str] = (),
-    llvm_kompile: bool = True,
-    debug: bool = False,
-    verbose: bool = False,
-) -> None:
-    from kevm_pyk.kompile import KompileTarget, kevm_kompile
-
-    syntax_module = 'FOUNDRY-CONTRACTS'
-    foundry = Foundry(foundry_root)
-    foundry_requires_dir = foundry.kompiled / 'requires'
-    foundry_contracts_file = foundry.kompiled / 'contracts.k'
-    kompiled_timestamp = foundry.kompiled / 'timestamp'
-    main_module = 'FOUNDRY-MAIN'
-    ensure_dir_path(foundry.kompiled)
-    ensure_dir_path(foundry_requires_dir)
-
-    requires_paths: dict[str, str] = {}
-
-    foundry.build()
-
-    if not foundry.up_to_date():
-        _LOGGER.info('Detected updates to contracts, regenerating K definition.')
-        regen = True
-
-    for r in requires:
-        req = Path(r)
-        if not req.exists():
-            raise ValueError(f'No such file: {req}')
-        if req.name in requires_paths.keys():
-            raise ValueError(
-                f'Required K files have conflicting names: {r} and {requires_paths[req.name]}. Consider changing the name of one of these files.'
-            )
-        requires_paths[req.name] = r
-        req_path = foundry_requires_dir / req.name
-        if regen or not req_path.exists():
-            _LOGGER.info(f'Copying requires path: {req} -> {req_path}')
-            shutil.copy(req, req_path)
-            regen = True
-
-    _imports: dict[str, list[str]] = {contract.name: [] for contract in foundry.contracts.values()}
-    for i in imports:
-        imp = i.split(':')
-        if not len(imp) == 2:
-            raise ValueError(f'module imports must be of the form "[ContractName]:[MODULE-NAME]". Got: {i}')
-        if imp[0] in _imports:
-            _imports[imp[0]].append(imp[1])
-        else:
-            raise ValueError(f'Could not find contract: {imp[0]}')
-
-    if regen or not foundry_contracts_file.exists() or not foundry.main_file.exists():
-        copied_requires = []
-        copied_requires += [f'requires/{name}' for name in list(requires_paths.keys())]
-        imports = ['FOUNDRY']
-        kevm = KEVM(DistTarget.FOUNDRY.get())
-        empty_config = kevm.definition.empty_config(Foundry.Sorts.FOUNDRY_CELL)
-        bin_runtime_definition = _foundry_to_contract_def(
-            empty_config=empty_config,
-            contracts=foundry.contracts.values(),
-            requires=['foundry.md'],
-        )
-
-        contract_main_definition = _foundry_to_main_def(
-            main_module=main_module,
-            empty_config=empty_config,
-            contracts=foundry.contracts.values(),
-            requires=(['contracts.k'] + copied_requires),
-            imports=_imports,
-        )
-
-        kevm = KEVM(
-            DistTarget.FOUNDRY.get(),
-            extra_unparsing_modules=(bin_runtime_definition.all_modules + contract_main_definition.all_modules),
-        )
-        foundry_contracts_file.write_text(kevm.pretty_print(bin_runtime_definition, unalias=False) + '\n')
-        _LOGGER.info(f'Wrote file: {foundry_contracts_file}')
-        foundry.main_file.write_text(kevm.pretty_print(contract_main_definition) + '\n')
-        _LOGGER.info(f'Wrote file: {foundry.main_file}')
-
-    def kompilation_digest() -> str:
-        k_files = list(requires) + [foundry_contracts_file, foundry.main_file]
-        return hash_str(''.join([hash_str(Path(k_file).read_text()) for k_file in k_files]))
-
-    def kompilation_up_to_date() -> bool:
-        if not foundry.digest_file.exists():
-            return False
-        digest_dict = json.loads(foundry.digest_file.read_text())
-        if 'kompilation' not in digest_dict:
-            digest_dict['kompilation'] = ''
-        foundry.digest_file.write_text(json.dumps(digest_dict, indent=4))
-        return digest_dict['kompilation'] == kompilation_digest()
-
-    def update_kompilation_digest() -> None:
-        digest_dict = {}
-        if foundry.digest_file.exists():
-            digest_dict = json.loads(foundry.digest_file.read_text())
-        digest_dict['kompilation'] = kompilation_digest()
-        foundry.digest_file.write_text(json.dumps(digest_dict, indent=4))
-
-        _LOGGER.info('Updated Kompilation digest')
-
-    if not kompilation_up_to_date() or rekompile or not kompiled_timestamp.exists():
-        kevm_kompile(
-            target=KompileTarget.HASKELL_BOOSTER,
-            output_dir=foundry.kompiled,
-            main_file=foundry.main_file,
-            main_module=main_module,
-            syntax_module=syntax_module,
-            includes=[include for include in includes if Path(include).exists()],
-            emit_json=True,
-            ccopts=ccopts,
-            llvm_library=foundry.llvm_library,
-            debug=debug,
-            verbose=verbose,
-        )
-
-    update_kompilation_digest()
-    foundry.update_digest()
-
-
 def foundry_show(
     foundry_root: Path,
     test: str,
@@ -905,46 +777,6 @@ def foundry_get_model(
 def _write_cfg(cfg: KCFG, path: Path) -> None:
     path.write_text(cfg.to_json())
     _LOGGER.info(f'Updated CFG file: {path}')
-
-
-def _foundry_to_contract_def(
-    empty_config: KInner,
-    contracts: Iterable[Contract],
-    requires: Iterable[str],
-) -> KDefinition:
-    modules = [contract_to_main_module(contract, empty_config, imports=['FOUNDRY']) for contract in contracts]
-    # First module is chosen as main module arbitrarily, since the contract definition is just a set of
-    # contract modules.
-    main_module = Contract.contract_to_module_name(list(contracts)[0].name_upper)
-
-    return KDefinition(
-        main_module,
-        modules,
-        requires=(KRequire(req) for req in list(requires)),
-    )
-
-
-def _foundry_to_main_def(
-    main_module: str,
-    contracts: Iterable[Contract],
-    empty_config: KInner,
-    requires: Iterable[str],
-    imports: dict[str, list[str]],
-) -> KDefinition:
-    modules = [
-        contract_to_verification_module(contract, empty_config, imports=imports[contract.name])
-        for contract in contracts
-    ]
-    _main_module = KFlatModule(
-        main_module,
-        imports=(KImport(mname) for mname in [_m.name for _m in modules]),
-    )
-
-    return KDefinition(
-        main_module,
-        [_main_module] + modules,
-        requires=(KRequire(req) for req in list(requires)),
-    )
 
 
 class FoundryNodePrinter(KEVMNodePrinter):
